r53: fixed a crash in unitig generation
When generating unitigs, forward and reverse strands may be of different
lengths. In rare cases, the length of the forward strand can be shorter than
the overlap length, which leads to integer underflow and subsequently a
segfault. There are no perfect solutions to this. This commit just makes the
overlap length shorter when this happens.